+++ DiceBot/Core/AppHelpers.cs
@@ -4,7 +4,7 @@
     class AppHelpers
     {
         public const string AppName = "DICEBot by WinMachine";
-        public const string AppVersion = "4.1.16.0";
+        public const string AppVersion = "4.1.18.5";
 
         public static string AppFullName
         {
