+++ DiceBot/Core/Shema/Bet.cs
@@ -1,23 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 namespace DiceBot
 {
-
-
-    class cBet
-    {
-        public string BetNr { get; set; }
-        public string Bet_Amount { get; set; }
-        public string Total_Wagered { get; set; }
-        public string Return_on_win { get; set; }
-        public string Profit { get; set; }
-    }
-
-
-
     public class Bet
     {
         public string Guid { get; set; }
