+++ DiceBot/Core/Shema/cBet.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace DiceBot
+{
+
+
+    class cBet
+    {
+        public string BetNr { get; set; }
+        public string Bet_Amount { get; set; }
+        public string Total_Wagered { get; set; }
+        public string Return_on_win { get; set; }
+        public string Profit { get; set; }
+    }
+
+}
+
