+++ DiceBot/Sites/bitsler/Bitlser.cs
