+++ DiceBot/Sites/stake/StakeOld.cs
